Implement real time logging
A new logging framework is implemented. All logging of import and migration
is done against the new `IMigrationLog` interface.

Two loggers are implemented:

- `HttpResponseLog` writes log messages to the HTTP response stream
- `Log4NetLog` wraps a log4net `ILog`

The migration page now uses AJAX to invoke the import and migration tasks.
This way we can easily listen to the Http log via the XHR's
onreadystatechange event and update the log window on the webpage in a
real time fashion.